LPS-94301 chat - classes in module that are not exported should be in internal package
